+++ CHANGES.rst
@@ -12,6 +12,8 @@ Version 0.2.0
 - Refactored the plugin core. There are two classes now: ``SingleSelectLisbox``
   and ``MultiSelectListbox``, which are inherited from the ``Listbox`` class.
 - Fixed #1: use selection of the parent element during initialization.
+- ``multiselect`` option doesn't used now.  Use ``multiple`` attribute of
+  the ``<select>`` tag instead.
 
 
 Version 0.1.3